* working on dynamic command execution for the webUI. works with a list of whitelisted commands which pipe the output into 'cache/cmd_runner.txt'

+++ include/transmission-install.sh
@@ -4,13 +4,21 @@ LANG=en_US.UTF-8
 export LANG
 
 # update system to latest first since transmission may depend on a latest lib
+echo "running system update first .... this may take a bit" > /usr/local/pia/cache/cmd_runner.txt
 /usr/local/pia/system-update.sh
 
-apt-get install -y transmission-cli transmission-daemon cifs-utils
+#apt-get install -y transmission-cli transmission-daemon cifs-utils
+pkg install transmission transmission-cli transmission-daemon >> /usr/local/pia/cache/cmd_runner.txt 2>> /usr/local/pia/cache/cmd_runner.txt
+if [ $? -ne 0 ]; then
+  echo 'fatal error while installing transmission. please install by hand' >> /usr/local/pia/cache/cmd_runner.txt
+  exit 1
+fi
+
 
 # make sure it is not running yet
 killall transmission-daemon
-update-rc.d transmission-daemon remove
+#update-rc.d transmission-daemon remove
 
 
-reboot
+echo 'transmission installed - rebooting system' >> /usr/local/pia/cache/cmd_runner.txt
+/sbin/reboot
